Licence_plate/requirements.txt
landingai
pandas
openpyxl
gdown
pillow
opencv-python
ipython